backend/requirements.txt
blinker==1.6.2
click==8.1.3
colorama==0.4.6
exceptiongroup==1.1.1
Flask==2.3.2
importlib-metadata==6.6.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.1
pluggy==1.0.0
pytest==7.3.1
python-dotenv==1.0.0
tomli==2.0.1
Werkzeug==2.3.4
zipp==3.15.0
